Enhance OBD telemetry support by updating protocol version and adding new PIDs: introduce fuel pressure, intake map, timing advance, turbo pressure, and turbo RPM to the telemetry structure. Update relevant functions and UI components to handle new data fields and ensure proper display in the gauge screens. Adjust valid_mask handling for new PIDs and refactor related code for improved clarity and functionality.

+++ lib/obd/service/collector.cpp
@@ -22,7 +22,7 @@ bool ObdCollector::poll(uint32_t now_ms) {
     last_poll_ms_ = now_ms;
     ++poll_count_;
 
-    uint8_t valid_mask = 0;
+    uint16_t valid_mask = 0;
     for (const uint8_t pid : kPidList) {
         ObdPidRaw raw{};
         if (obd_.readPid(pid, raw) != ObdError::Ok) {
